requirements.txt
PyQt5==5.15.10
mysql-connector-python==8.2.0
python-dotenv==1.0.0
bcrypt==4.1.2
reportlab==4.0.9
openpyxl==3.1.2
Pillow==10.2.0